renaming vars and code cleanup after review
